_All_Projects/12.8/requirements.txt
Django>=3,<4
djangorestframework
